fix(catalog): correct spelling and enhance delete confirmation for products
Updated the spelling of "Delete bran" to "Delete brand" in the brand update component. Improved user experience by adding confirmation alerts before deleting products in both the product card and update product components. Adjusted formatting in the all products component for better readability.

+++ src/app/pages/catalog/product/update-product/update-product.component.ts
@@ -1314,19 +1314,23 @@ export class UpdateProductComponent implements OnInit {
   }
 
   onDelete() {
-    this.ProductService.deleteProduct(this.productDetails?._id).subscribe({
-      next: (res: any) => {
-        if (res?.errorCode == 0) {
-          this.HotToastService.success(res?.message);
-          this.Router.navigate(['/app/product']);
-        } else {
-          this.HotToastService.error(res?.message);
-        }
-      },
-      error: (err: any) => {
-        this.HotToastService.error(err.error.message);
-      },
-    });
+    if (confirm('Are you sure you want to delete this product?')) {
+      this.ProductService.deleteProduct(this.productDetails?._id).subscribe({
+        next: (res: any) => {
+          if (res?.errorCode == 0) {
+            this.HotToastService.success(res?.message);
+            this.Router.navigate(['/app/product']);
+          } else {
+            this.HotToastService.error(res?.message);
+          }
+        },
+        error: (err: any) => {
+          this.HotToastService.error(err.error.message);
+        },
+      });
+    } else {
+      this.HotToastService.error('Action cancelled')
+    }
   }
 
   handleTagIcons(event: any) {
